temporal_workflow/requirements.txt
temporalio==1.15.0
pymongo==4.12.0
python-dotenv==1.1.0